# go/model_student.go
/*
 * Телеграм Бот для работы с научными группами
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 0.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type Student struct {
	// Id студента
	StudentId int32 `json:"student_id,omitempty"`
	// Id его научного руководителя
	TutorId int32 `json:"tutor_id,omitempty"`
	// Имя студента
	Name string `json:"name"`
	// Фамилия студента
	Surname string `json:"surname"`
	// Курс студента
	Course int32 `json:"course,omitempty"`
	// Тема курсовой работы студента
	CourseWorkSubject string `json:"course_work_subject,omitempty"`
	// Тема дипломной работы студента
	DiplomaWorkSublect string `json:"diploma_work_sublect,omitempty"`
	// Id как пользователя
	UserId int32 `json:"user_id"`
}