# go/model_user.go
/*
 * Телеграм Бот для работы с научными группами
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 0.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type User struct {
	// Id пользователя
	UserId int32 `json:"user_id"`
	// Id в ролевой системе (студент или преподователь)
	RoleId int32 `json:"role_id"`
	// Hash пароля
	PasswordHash string `json:"password_hash"`
	// Соль пароля
	Salt string `json:"salt"`
}